requirements.txt
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.2.0
langchain-text-splitters==0.2.0
langchain-ollama==0.2.0
huggingface-hub
faiss-cpu
pydub
sounddevice
numpy
torch
transformers
openai-whisper
silero-vad
simpleaudio
soundfile
unstructured
coqui-tts